changed: put ratings videodatabase methods in separate file
